removed abstract organization class
