+++ src/SWP/Component/MultiTenancy/Model/Organization.php
@@ -15,9 +15,29 @@ namespace SWP\Component\MultiTenancy\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use SWP\Component\Common\Model\EnableableTrait;
+use SWP\Component\Common\Model\SoftDeletableTrait;
+use SWP\Component\Common\Model\TimestampableTrait;
 
-class Organization extends AbstractOrganization implements OrganizationInterface
+class Organization implements OrganizationInterface
 {
+    use SoftDeletableTrait, TimestampableTrait, EnableableTrait;
+
+    /**
+     * @var mixed
+     */
+    protected $id;
+
+    /**
+     * @var string
+     */
+    protected $name;
+
+    /**
+     * @var string
+     */
+    protected $code;
+
     /**
      * @var Collection|TenantInterface[]
      */
@@ -30,6 +50,58 @@ class Organization extends AbstractOrganization implements OrganizationInterface
     {
         $this->createdAt = new \DateTime();
         $this->tenants = new ArrayCollection();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setId($id)
+    {
+        $this->id = $id;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setName($name)
+    {
+        $this->name = $name;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getCode()
+    {
+        return $this->code;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setCode($code)
+    {
+        if (null !== $this->code) {
+            throw new \LogicException('The code is already set. Can not change it.');
+        }
+
+        $this->code = $code;
     }
 
     /**
